If no colors are specified, use plain fmt.Sprintf func
If no colors are specified in the `Config` struct, we should not apply color
formatting and assume the user would give us a set of colors if they wanted to
use them. This removes the default `fgWhite` setting for Config.Colors in
`New()`, and removes the setting of Config.StopColors to the value of
Config.Colors if StopColors hadn't been provided.