Add StopFail() method to Spinner for printing a failure message
This change adds a `StopFail()` method, to compliment the `Stop()` method, and
adds all the associated `Config` fields and configuration methods. The purpose
of this functionality is to allow users to treat `Stop()` like a success result,
and `StopFail()` as a failure result. They can provide different characters,
colors, and messages for each.

One use case for this would be to provide the error to an end user if the task
were to fail, but to provide no message if it were to succeed.