Make Spinner internal cancel (stop) channel buffered
The original implementation of this channel used only closing to signal, which
is a non-blocking operation. When adding `StopFail()`, the `close()` became the
signal for `StopFail()` and sending an empty struct over the channel became the
signal for `Stop()`. We don't want to block the channel send on the painter
being in a position to receive the value, so we make the channel buffered with a
size of 1.

This also adds a comment on the cancelCh Spinner field, briefly noting the
semantics of its use.