Remove sync/atomic.Value in favor of mutex, add Stop*() tests
Atomic values made sense when the original implementation was simpler, but as it
grew the addtional features were wedged-in even when atomic values no longer
made sense. This was because the values were related to other fields on the
struct so they couldn't be truly atomic.

Let's instead use a mutex so spinner paints are consistent, and so the
implementationn is simpler.

Finally, this adds tests for `Stop()` and `StopFail()`, which both cover the
internal `stop()` method.