Add support for Windows printing
This adds support for printing on Windows, using a more rudimentary method for
clearing the screen. Windows terminals don't understand VT100 / ANSI terminal
escape sequences, so the methods used for OS X and Linux aren't applicable.

When on Windows, it tracks the number of wrtten bytes and on erase writes a
carriage return and that number of plain spaces to overwrite the characters
followed by a final carriage return.