Separate updates to content from spinner animation
Prior to this, your spinner content could only be updated as often as the
spinner was rendering. Data liveliness was tied to your animation speed, which
meant for faster updates you had to sacrifice what your spinner animation looked
like.

This change separates the two, so that calling a method like `Message()` should
update the spinner's message in near realtime without the spinner itself being
animated.