Add ability to pause spinner and get its status
This change adds three new methods to the `Spinner` type, and deprecates the
`Active` method.

The first are the `Pause()` and `Unpause()` methods, which put the spinner in a
paused state so that you can update multiple configuration options without users
seeing a spinner rendered with partial changes applied. Such as changing the
suffix, message, or colors.

The third method is the `Status()` method, which deprecates `Active()`. Because
the internal state machine for the spinner has become more complex, it's no
longer reasonable to consider it active or inactive. With this we include a new
type, `SpinnerStatus`, and constants to represent each state of the internal
state machine:

- SpinnerStopped
- SpinnerStarting
- SpinnerRunning
- SpinnerStoppping
- SpinnerPausing
- SpinnerPaused
- SpinnerUnpausing