Reduce missed Frequency updates, reduce mutex locks, update docs
This change is a small update to help make the spinner a bit more robust.

The first increases internal channel buffer size for frequency updates, to hopefully
non-blocking write dropping an update. The buffer size increase is arbitrary,
and calling .Frequency() in a hot loop could still result in dropped updates.

The second is the removal of a mutex lock within the Status() method, and an
unnecessary nil check. There is not a need for either of them based on how
the *Spinner is constructed.

The third is some refinemennt of the documentation both in comments on exported
identifiers, and comments within the source code.