Use a byte buffer to reduce number of write calls to stdout
This change introduces a 2KB (growable) byte buffer, to sit in between the
painting function(s) and stdout. Previously, as we were animating the spinner we
would write the bytes directly to stdout as we were generating the animation.

This change makes it so that most of our writes are to an in-memory buffer, and
then we only have one system call to write to the actaul terminal the contents
of that buffer.

This should act like a bit of a frame buffer, so if if anyone was experiencing
any sort of stutter, where they could see the cursor as the spinner was
animated, I believe this should result in a smoother adnimation.