Add support for non-TTY output targets
This change introduces better handling of non-TTY output targets. Specifically,
the spinner animation is disabled and the lines aren't erased. In the Non-TTY
mode, the spinner is only animated on each update to the spinner data (e.g.,
message). In addition, this does not erase the line when rendering and instead
renders the animation on a new line.