Import more CharSets and license header from briandowns/spinner
The project where we've gotten the spinners from has added 10 more of them, and
added a new license header to the file we copied them from. So we import both
the new spinners and the new copyright header, while adding my own copyright for
the delcared modifications made to the file.