Make CharSets test less brittle to future changes
When `yacspin` adds its own pre-created spinners to the `CharSets` variable,
they will start at index 500 to ideally avoid conflicts with spinners pulled-in
from github.com/briandowns/spinner. However, if that change was made today this
test would likely fail.

So let's update the test to range over the map directly, instead of using
indexes, to avoid this breaking in the future.